restructured file tree, placed components and composables inside of UI package folder.
