--- app/src/main/java/com/ls/localsky/app/LocalSkyApp.kt
@@ -1,19 +1,0 @@
-package com.ls.localsky.app
-
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.Surface
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-
-@Composable
-fun LocalSkyApp(){
-
-    Surface(
-        modifier = Modifier.fillMaxSize(),
-        color = Color.White
-    ){
-
-    }
-
-}
