+++ app/src/main/java/com/ls/localsky/ui/app/LocalSkyApp.kt
@@ -1,0 +1,33 @@
+package com.ls.localsky.ui.app
+
+import androidx.compose.animation.Crossfade
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import com.ls.localsky.ui.screens.LoginScreen
+import com.ls.localsky.ui.screens.RegisterScreen
+
+@Composable
+fun LocalSkyApp(){
+
+    Surface(
+        modifier = Modifier.fillMaxSize(),
+        color = Color.White
+    ) {
+
+        Crossfade(targetState = LocalSkyAppRouter.currentScreen, label = "") { currentState ->
+            when(currentState.value) {
+                is Screen.LoginScreen -> {
+                    LoginScreen()
+                }
+                is Screen.RegisterScreen -> {
+                    RegisterScreen()
+                }
+            }
+        }
+
+    }
+
+}
