+++ app/src/main/java/com/ls/localsky/ui/screens/LoginScreen.kt
@@ -1,4 +1,4 @@
-package com.ls.localsky.loginScreen
+package com.ls.localsky.ui.screens
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
@@ -14,13 +14,14 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.ls.localsky.R
-import com.ls.localsky.components.ButtonComponent
-import com.ls.localsky.components.ClickableLoginText
-import com.ls.localsky.components.ClickableRegisterText
-import com.ls.localsky.components.DividerTextComponent
-import com.ls.localsky.components.EmailInput
-import com.ls.localsky.components.PasswordInput
-import com.ls.localsky.components.TitleText
+import com.ls.localsky.ui.app.LocalSkyAppRouter
+import com.ls.localsky.ui.app.Screen
+import com.ls.localsky.ui.components.ButtonComponent
+import com.ls.localsky.ui.components.ClickableRegisterText
+import com.ls.localsky.ui.components.DividerTextComponent
+import com.ls.localsky.ui.components.EmailInput
+import com.ls.localsky.ui.components.PasswordInput
+import com.ls.localsky.ui.components.TitleText
 
 @Composable
 fun LoginScreen() {
@@ -45,7 +46,9 @@ fun LoginScreen() {
             Spacer(modifier=Modifier.height(10.dp))
             //Login with google
             DividerTextComponent()
-            ClickableRegisterText(onTextSelected = {})
+            ClickableRegisterText(onTextSelected = {
+                LocalSkyAppRouter.navigateTo(Screen.RegisterScreen)
+            })
 
         }
